backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
SQLAlchemy==2.0.25
psycopg==3.1.18
pydantic==2.6.1
python-dateutil==2.9.0
PyJWT==2.8.0
python-dotenv==1.0.1
